# Conflicts:
#	src/main/java/ua/softserve/ita/config/SecurityConfiguration.java
#	src/main/java/ua/softserve/ita/controller/ClaimController.java
#	src/main/java/ua/softserve/ita/controller/SearchController.java
#	src/main/java/ua/softserve/ita/controller/VacancyController.java
#	src/main/java/ua/softserve/ita/dao/impl/PdfResumeDaoImpl.java
#	src/main/java/ua/softserve/ita/dao/impl/search/SearchResumeDao.java
#	src/main/java/ua/softserve/ita/dao/impl/search/SearchVacancyDao.java
#	src/main/java/ua/softserve/ita/dto/company/CompanyPaginationDto.java
#	src/main/java/ua/softserve/ita/dto/search/SearchResumeDto.java
#	src/main/java/ua/softserve/ita/model/Company.java
#	src/main/java/ua/softserve/ita/model/User.java
#	src/main/java/ua/softserve/ita/service/impl/ResumeServiceImpl.java
#	src/main/java/ua/softserve/ita/service/impl/search/SearchResumeService.java
#	src/main/java/ua/softserve/ita/service/impl/search/SearchVacancyService.java
#	src/main/java/ua/softserve/ita/service/mail/MailService.java
#	src/main/java/ua/softserve/ita/service/mail/MailServiceImpl.java
#	src/main/java/ua/softserve/ita/service/pdfcreator/CleanTempResumePdf.java
#	src/main/java/ua/softserve/ita/service/pdfcreator/CreateQrCodeVCard.java
#	src/main/java/ua/softserve/ita/service/pdfcreator/CreateResumePdf.java
#	src/main/java/ua/softserve/ita/service/search/SearchResumeMapper.java
#	src/main/java/ua/softserve/ita/service/search/SearchVacancyMapper.java
#	src/main/resources/log4j2.properties
#	src/test/java/ua.softserve.ita/dao/SearchResumeDaoTest.java

+++ src/main/java/ua/softserve/ita/dto/search/SearchResumeDto.java
@@ -1,13 +1,12 @@
 package ua.softserve.ita.dto.search;
 
-import lombok.Builder;
-import lombok.Data;
+import lombok.*;
 
 @Data
 @Builder
 public class SearchResumeDto {
 
-    private long id;
+    private BigInteger id;
 
     private String firstName;
 
@@ -17,7 +16,7 @@ public class SearchResumeDto {
 
     private String position;
 
-    private long resumeId;
+    private BigInteger resumeId;
 
     private String city;
